+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2020-11-27
+### Added
+- Custom Upgradable Proxy contract that behaves similarly to the [EIP-1822](https://eips.ethereum.org/EIPS/eip-1822): Universal Upgradeable Proxy Standard (UUPS), except that it points to an Authority contract which in itself points to an implementation (which can be updated).
+- New ImplementationAuthority contract that acts as an authority for proxy contracts
+- Library Lock contract to ensure no one can manipulate the Logic Contract once it is deployed
+- Version contract that gives the versioning information of the implementation contract
+### Moved
+- variables in a separate contract (Storage.sol)
+- structs in a separate contract (Structs.sol)
+### Updated
+- Update contracts to SOL =0.6.9
+
 ## [1.1.2] - 2020-09-30
 ### Fixed
 - Add Constructor on ClaimIssuer Contract